we should still fail if there's no metadata
Sometimes, all metadata is unsigned. We should tell people about that,
but it's not a show-stopper. If, however, we didn't generate any
metadata, we have a problem, and we should stop.

Because that's a "top level" error result (exit 1), it makes sense to
use other return codes for errors that are deeper in the code.